Moving vanity URL utility to PageUtils to follow same structure with AEM
